editing comments from Jiefei 022725
Please use window and step instead of tWindow and tStep. Also do a global search to make sure all tWindow and tStep can be changed
elecSoz to sozIndex. Also global search. Please remove "(for good electrodes)" as it does not have to be the good one.
timeWindow to timeRange to avoid the possible confusion between window and timeWindow